move loading progress from pages to components
